Bugfix: in resolver mechanism
If resolver does a reverse DNS lookup and times out, requests in the
same resolver node's waiting list will be blocked forever and nginx will
not be reloaded or shuted down gracefully.

This fixup modifies the timeout handler of resolver module to go
through all the remaining ctx in rn->waiting and calls each ctx's
handler to do the cleanup work.

Signed-off-by: Paul Yang(InfoHunter) <paulyang.inf@gmail.com>