Feature: ssl client/server certificate in upstream
This feature introduces the ssl client certificate authentication
functionality when proxy requests to a https backend server which
requires a client certificate authentication.

You can specify a cert and a key to use when handshaking with backend
https servers, the cert will be used by the server to authenticate.

You can also turn on server certificate authentication at SEnginx side
to verify the server's cert, in this case, you also need to specify the
ca cert that backend server uses.

5 new directives in ngx_http_proxy_module are introduced:
 * proxy_ssl_certificate
 * proxy_ssl_certificate_key
 * proxy_ssl_verify_server
 * proxy_ssl_verify_depth
 * proxy_ssl_server_certificate

This feature can solve the following requirement:
http://forum.nginx.org/read.php?2,221563,221563#msg-221563

Signed-off-by: paulyang <y_y@neusoft.com>