Bugfix: issue #20
Fixed issue #20, which reported that if keepalive is enabled in an
upstream and dynamic resovle is also enabled, SEnginx worker will die due to
a segmentation fault caused by a "double free" error.

Fixed as:

1) use NGX_STOP to indicate to pause current request to wait for a DNS
query response. NGX_DONE is used by the upstream keepalive module to
connect to upstream server.

2) introduce new test cases to cover this scenario.

Signed-off-by: Paul Yang(InfoHunter) <paulyang.inf@gmail.com>