Feature: add robot_mitigation_secret directive
Use this directive to set a secret string to enhance robustness of the
challenge key, if this direcive is missing, SEnginx will pick a dynamic
secret each time on start. You can also use this directive to set a same
secret among some SEnginxes in a cluster to avoid "miss-match" problem
of robot mitigation.

Signed-off-by: Paul Yang(InfoHunter) <paulyang.inf@gmail.com>